+++ release/index.js
@@ -78,7 +78,7 @@ return /******/ (function(modules) { // webpackBootstrap
 /***/ "./package.json":
 /***/ (function(module, exports) {
 
-module.exports = {"name":"ngx-uploadcare-widget","version":"1.0.1","description":"Angular 2+ wrapper for Uploadcare Widget","main":"release/index.js","typings":"release/index.d.ts","scripts":{"check":"npm-check --skip-unused","test":"npm run lint && karma start","test:watch":"npm run test -- --auto-watch --no-single-run","watch:test":"npm run test -- --auto-watch --no-single-run","lint":"tslint --project tsconfig.json --config tslint.json","version":"npm run package","preversion":"npm test","clean":"npm-run-all -p clean:*","clean:build":"rimraf build","clean:dist":"rimraf dist","clean:release":"rimraf release","build":"webpack --display-error-details","build:release":"cross-env NODE_ENV=production npm run build","build:package":"cross-env NODE_ENV=package npm run build","build:ts":"tsc","build:aot":"ngc","build:sass":"node-sass -r build/ -o build/","build:css":"postcss --use autoprefixer dist/*.css -d dist/","copy:build":"cpx \"src/**/*\" build","copy:styles":"cpx \"build/**/*.css\" release","watch":"webpack --display-error-details --watch","start":"webpack-dev-server","start:host":"webpack-dev-server --host 0.0.0.0","start:hmr":"webpack-dev-server --env.HMR","release":"npm-run-all build:release","package":"npm-run-all -s clean copy:build package:aot build:package package:minify clean:build","package:ts":"tsc --outDir release","package:aot":"ngc -p tsconfig-aot.json","package:replace-scss":"node ./config/replace-scss.js","package:minify":"uglifyjs release/index.js --source-map --compress --mangle --screw-ie8 --output release/index.min.js","profile":"NODE_ENV=production webpack --profile --json > dist/stats.json"},"repository":{"type":"git","url":"git+https://github.com/uploadcare/ngx-uploadcare-widget.git"},"keywords":["angular","ngx","uploadcare","widget","uploader","filepicker","dialog","upload","store","file","files","image","images"],"authors":["Nikolay Zherdev"],"license":"MIT","bugs":{"url":"https://github.com/uploadcare/ngx-uploadcare-widget/issues"},"homepage":"https://github.com/uploadcare/ngx-uploadcare-widget#readme","peerDependencies":{"@angular/common":"5.x","@angular/core":"5.x","@angular/animations":"5.x","@angular/platform-browser":"5.x","rxjs":"5.x"},"devDependencies":{"@angular/animations":"^5.2.3","@angular/common":"^5.2.3","@angular/compiler":"^5.2.3","@angular/compiler-cli":"^5.2.3","@angular/core":"^5.2.3","@angular/forms":"^5.2.3","@angular/http":"^5.2.3","@angular/platform-browser":"^5.2.3","@angular/platform-browser-dynamic":"^5.2.3","@angular/platform-server":"^5.2.3","@swimlane/ngx-ui":"^19.2.0","@types/jasmine":"^2.8.6","@types/node":"^8.5.10","angular2-prettyjson":"^3.0.1","angular2-template-loader":"^0.6.0","autoprefixer":"^7.2.5","awesome-typescript-loader":"^3.2.3","bootstrap":"^4.0.0","clean-webpack-plugin":"^0.1.18","codelyzer":"^3.0.0-beta.4","copy-webpack-plugin":"^4.0.0","core-js":"^2.5.1","cpx":"^1.5.0","cross-env":"^5.0.5","css-loader":"^0.28.9","emoji-flags":"^1.2.0","extract-text-webpack-plugin":"2.0.0-beta.4","file-loader":"^0.11.2","flexboxgrid":"^6.3.1","fs-extra":"^4.0.1","gh-pages":"^1.0.0","html-loader":"^0.5.5","html-webpack-plugin":"^2.26.0","istanbul-instrumenter-loader":"^3.0.0","jasmine-core":"^2.99.0","jquery":"^3.3.1","karma":"^1.7.1","karma-chrome-launcher":"^2.0.0","karma-coverage":"^1.1.1","karma-jasmine":"^1.1.0","karma-mocha-reporter":"^2.2.4","karma-remap-coverage":"^0.1.5","karma-sourcemap-loader":"^0.3.7","karma-webpack":"^2.0.1","node-sass":"^4.3.0","npm-run-all":"^4.1.1","popper.js":"^1.13.0","postcss":"^5.2.11","postcss-loader":"^1.2.2","raw-loader":"^0.5.1","replace":"^0.3.0","rimraf":"^2.5.2","rxjs":"^5.5.2","sass-loader":"^6.0.3","source-map-loader":"^0.2.0","style-loader":"^0.18.2","to-string-loader":"^1.1.5","ts-helpers":"1.1.2","tslint":"^5.9.1","tslint-config-swimlane":"^3.0.3","tslint-loader":"^3.3.0","typescript":"2.5.2","uglify-js":"^3.3.9","url-loader":"^0.5.7","web-animations-js":"^2.2.2","webpack":"^3.5.6","webpack-combine-loaders":"^2.0.3","webpack-dev-server":"^2.11.1","webpack-merge":"^4.1.0","webpack-notifier":"^1.5.1","zone.js":"^0.8.20"},"dependencies":{"@angular/core":"^5.2.3","@angular/platform-browser":"^5.2.3","uploadcare-widget":"3.x"}}
+module.exports = {"name":"ngx-uploadcare-widget","version":"1.0.1","description":"Angular 2+ wrapper for Uploadcare Widget","main":"release/index.js","typings":"release/index.d.ts","scripts":{"check":"npm-check --skip-unused","test":"npm run lint && karma start","test:watch":"npm run test -- --auto-watch --no-single-run","watch:test":"npm run test -- --auto-watch --no-single-run","lint":"tslint --project tsconfig.json --config tslint.json","version":"npm run package","preversion":"npm test","clean":"npm-run-all -p clean:*","clean:build":"rimraf build","clean:dist":"rimraf dist","clean:release":"rimraf release","build":"webpack --display-error-details","build:release":"cross-env NODE_ENV=production npm run build","build:package":"cross-env NODE_ENV=package npm run build","build:ts":"tsc","build:aot":"ngc","build:sass":"node-sass -r build/ -o build/","build:css":"postcss --use autoprefixer dist/*.css -d dist/","copy:build":"cpx \"src/**/*\" build","copy:styles":"cpx \"build/**/*.css\" release","watch":"webpack --display-error-details --watch","start":"webpack-dev-server","start:host":"webpack-dev-server --host 0.0.0.0","start:hmr":"webpack-dev-server --env.HMR","release":"npm-run-all build:release","package":"npm-run-all -s clean copy:build package:aot build:package package:minify clean:build","package:ts":"tsc --outDir release","package:aot":"ngc -p tsconfig-aot.json","package:replace-scss":"node ./config/replace-scss.js","package:minify":"uglifyjs release/index.js --source-map --compress --mangle --screw-ie8 --output release/index.min.js","profile":"NODE_ENV=production webpack --profile --json > dist/stats.json"},"repository":{"type":"git","url":"git+https://github.com/uploadcare/ngx-uploadcare-widget.git"},"keywords":["angular","ngx","uploadcare","widget","uploader","filepicker","dialog","upload","store","file","files","image","images"],"authors":["Nikolay Zherdev"],"license":"MIT","bugs":{"url":"https://github.com/uploadcare/ngx-uploadcare-widget/issues"},"homepage":"https://github.com/uploadcare/ngx-uploadcare-widget#readme","peerDependencies":{"@angular/common":"5.x","@angular/core":"5.x","@angular/animations":"5.x","@angular/platform-browser":"5.x","rxjs":"5.x"},"devDependencies":{"@angular/animations":"^5.2.11","@angular/common":"^5.2.11","@angular/compiler":"^5.2.11","@angular/compiler-cli":"^5.2.11","@angular/core":"^5.2.11","@angular/forms":"^5.2.11","@angular/http":"^5.2.11","@angular/platform-browser":"^5.2.11","@angular/platform-browser-dynamic":"^5.2.11","@angular/platform-server":"^5.2.11","@swimlane/ngx-ui":"^19.2.0","@types/jasmine":"^2.8.8","@types/node":"^8.10.21","angular2-prettyjson":"^3.0.1","angular2-template-loader":"^0.6.0","autoprefixer":"^7.2.6","awesome-typescript-loader":"^3.5.0","bootstrap":"^4.1.3","clean-webpack-plugin":"^0.1.19","codelyzer":"^3.0.0-beta.4","copy-webpack-plugin":"^4.5.2","core-js":"^2.5.7","cpx":"^1.5.0","cross-env":"^5.2.0","css-loader":"^0.28.11","emoji-flags":"^1.2.0","extract-text-webpack-plugin":"2.0.0-beta.4","file-loader":"^0.11.2","flexboxgrid":"^6.3.1","fs-extra":"^4.0.1","gh-pages":"^1.2.0","html-loader":"^0.5.5","html-webpack-plugin":"^2.26.0","istanbul-instrumenter-loader":"^3.0.1","jasmine-core":"^2.99.1","jquery":"^3.3.1","karma":"^1.7.1","karma-chrome-launcher":"^2.0.0","karma-coverage":"^1.1.2","karma-jasmine":"^1.1.2","karma-mocha-reporter":"^2.2.4","karma-remap-coverage":"^0.1.5","karma-sourcemap-loader":"^0.3.7","karma-webpack":"^2.0.13","node-sass":"^4.9.2","npm-run-all":"^4.1.3","popper.js":"^1.14.3","postcss":"^5.2.11","postcss-loader":"^1.2.2","raw-loader":"^0.5.1","replace":"^0.3.0","rimraf":"^2.5.2","rxjs":"^5.5.11","sass-loader":"^6.0.7","source-map-loader":"^0.2.0","style-loader":"^0.18.2","to-string-loader":"^1.1.5","ts-helpers":"1.1.2","tslint":"^5.11.0","tslint-config-swimlane":"^3.0.3","tslint-loader":"^3.6.0","typescript":"2.5.2","uglify-js":"^3.4.5","url-loader":"^0.5.7","web-animations-js":"^2.2.2","webpack":"^3.12.0","webpack-combine-loaders":"^2.0.3","webpack-dev-server":"^2.11.2","webpack-merge":"^4.1.3","webpack-notifier":"^1.6.0","zone.js":"^0.8.26"},"dependencies":{"uploadcare-widget":"^3.5.1"}}
 
 /***/ }),
 
@@ -149,6 +149,7 @@ var UcWidgetComponent = /** @class */ (function () {
         this._cdnBase = null;
         this._reinitRequired = false;
         this._isClearValue = false;
+        this._validators = [];
         this.element = element;
         this.renderer = renderer;
     }
@@ -299,6 +300,15 @@ var UcWidgetComponent = /** @class */ (function () {
         enumerable: true,
         configurable: true
     });
+    Object.defineProperty(UcWidgetComponent.prototype, "validators", {
+        get: function () { return this._validators; },
+        set: function (validatorsArr) {
+            this._validators = validatorsArr;
+            this.setReinitFlag(false);
+        },
+        enumerable: true,
+        configurable: true
+    });
     Object.defineProperty(UcWidgetComponent.prototype, "cdnBase", {
         get: function () { return this._cdnBase; },
         set: function (cdnBase) {
@@ -387,6 +397,14 @@ var UcWidgetComponent = /** @class */ (function () {
         }
         this.initInputElement();
         var widget = __WEBPACK_IMPORTED_MODULE_1_uploadcare_widget___default.a.Widget(this.inputElement);
+        this._validators.forEach(function (v) {
+            if (typeof v === 'function') {
+                widget.validators.push(v);
+            }
+            else {
+                throw new Error('Only functions allowed in validadators array');
+            }
+        });
         widget.onUploadComplete(function (fileInfo) {
             _this.onUploadComplete.emit(fileInfo);
             _this._value = fileInfo.uuid;
@@ -510,6 +528,11 @@ var UcWidgetComponent = /** @class */ (function () {
         __metadata("design:paramtypes", [String])
     ], UcWidgetComponent.prototype, "value", null);
     __decorate([
+        Object(__WEBPACK_IMPORTED_MODULE_0__angular_core__["Input"])('validators'),
+        __metadata("design:type", Array),
+        __metadata("design:paramtypes", [Array])
+    ], UcWidgetComponent.prototype, "validators", null);
+    __decorate([
         Object(__WEBPACK_IMPORTED_MODULE_0__angular_core__["Input"])('cdn-base'),
         __metadata("design:type", String),
         __metadata("design:paramtypes", [String])
@@ -624,6 +647,7 @@ var UcWidgetCustomComponent = /** @class */ (function () {
         this.onChange = new __WEBPACK_IMPORTED_MODULE_0__angular_core__["EventEmitter"]();
         this.onProgress = new __WEBPACK_IMPORTED_MODULE_0__angular_core__["EventEmitter"]();
         this.publicKey = 'demopublickey';
+        this.validators = [];
     }
     UcWidgetCustomComponent.prototype.openDialog = function () {
         var _this = this;
@@ -645,7 +669,13 @@ var UcWidgetCustomComponent = /** @class */ (function () {
             secureExpire: this.secureExpire,
             value: this.value ? this.value : undefined,
             cdnBase: this.cdnBase ? this.cdnBase : undefined,
-            doNotStore: this.doNotStore
+            doNotStore: this.doNotStore,
+            validators: this.validators.filter(function (v) {
+                if (typeof v !== 'function') {
+                    throw new Error('Only functions allowed in validadators array');
+                }
+                return v;
+            }),
         };
         var dialog = __WEBPACK_IMPORTED_MODULE_1_uploadcare_widget___default.a.openDialog(this.value, null, config);
         dialog.done(function (selectionPromise) {
@@ -754,6 +784,10 @@ var UcWidgetCustomComponent = /** @class */ (function () {
         Object(__WEBPACK_IMPORTED_MODULE_0__angular_core__["Input"])('do-not-store'),
         __metadata("design:type", Boolean)
     ], UcWidgetCustomComponent.prototype, "doNotStore", void 0);
+    __decorate([
+        Object(__WEBPACK_IMPORTED_MODULE_0__angular_core__["Input"])('validators'),
+        __metadata("design:type", Array)
+    ], UcWidgetCustomComponent.prototype, "validators", void 0);
     UcWidgetCustomComponent = __decorate([
         Object(__WEBPACK_IMPORTED_MODULE_0__angular_core__["Component"])({
             selector: 'ngx-uploadcare-widget-custom',